Simple rewrite to Digest::HMAC helpers
We need to concretize algorithms to defined place instead of composition
in code. Digest::HMAC_* helpers are in the same package as Digest::HMAC.

Difference is that all code for algoritm is now in Digest::HMAC
distribution instead of previous situation when was composition
Digest::HMAC and Digest::SHA1 and Digest::MD5 in this code.